requirements.txt
anthropic>=0.18.0
feedparser>=6.0.11
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
sendgrid>=6.11.0
curl_cffi>=0.6.0
PyMuPDF>=1.23.0
websocket-client>=1.7.0
pip-system-certs>=4.0
